+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## dev
+### Fixed
+- Handle exception that occurs when backend init trys to find a deleted chat
+
 ## [0.4.0] - 2023.2.2
 
 ### Added